Add sequence number so sessions are doctested in file order.
The code generated for session Markdown fenced code blocks
is a def function_name(): followed by a docstring containing
the contents of the fenced code block.
A 5 digit sequence number is added to the function name
like this: def session_00001_line_6():.  The sequence number
is bumped for each session added to the generated pytest file.
This is needed so that doctest will run the sessions in file order.
Without this change doctest was observed to run session_10
before session_6.