Check callers setup block not using reserved name.
The name _session_globals is reserved for setting up
doctests.  It can't be used in a setup code block.

o add logic to fail if _session_globals is used in a setup block.
o add test case for _session_globals used in a setup block.
0 add test case to show setup_module()'s k, v can be used.
o fix teardown tests cases that fail on Python 3.5.