Bugfix- #4 --setup and --setup-doctest fails in pypy3
Refactor generated code that copies names to the module's namespace.
Pass the dict view returned by BIF locals() to two new functions
that do the iteration.  Since this passes the tests, it is likely
that the k, v in for k, v in locals.items() got into the dict view
during the iteration.  Possibly related or contributing to the
root cause of the issue is pytest assertion rewriting,
coverage code instrumentation, pypy3 JIT compilation, or just
the Python language implementation of k, v as local variables to be
included or not included in the locals() dict view.

Update tests to cover new code and handle the changes to the
examples of generated code in the documentation.